refactor: remove duplicate code
Removed duplication of formatting functions, which should
be part of useNetworkStore(), as they contain `currentNetwork`
(state of the app).

As a quickfix getEpochFromTimestamp() and getTimestampFromEpoch()
were `proped down` to DashboardChartTooltipHeader.vue
which is an `antipattern` and should be removed.

Reason: Due to the fact that DashboardChartTooltipHeader.vue was
used inside a `vue render()` in DashboardChartSummary.vue,
we do not have access to `nuxt hooks` so the
usage of useRuntimeConfig() in useNetworkStore() created an error.