refactor:  add `subpath` for server routes to avoid name collisions with `v1 app`
This change is necessary, since `v1 app`  and `v2 (this) app` are
running on the same `url`. So the traffic originating from the browser
to request a specific route to be routed to the
`v2 app instance` on the server.

That's why we use `/api/bff`, as this will be routed to this app
instance instead of `v1`.