refactor: move all assets to subfolder due to name collisions with `v1 app`
This change is necessary, since `v1 app`  and `v2 (this) app` are
running on the same `url`. So the traffic originating from the browser
to request a specific resource of this app has to be routed to the
`v2 app instance` on the server.

That's why we use a `generic prefix`(/assets-2usdf) for
`publicly served assets`.This will prevent name collisions.