fix: error for `empty dashboard` and `private dashboard id` in cookie
Before the `isGuestDashboardKey()` utility function threw an error,
in the `useDasboardKeyProvider()` composable.
Only when the url was `/dashboard` (no dashboard key) and the
`bc-validator-dashboard-key` cookie value was a number
(e.g. private dashboards).

Trying to use startsWith() on a number remained undetected since
the `useBcCookie() composable` used `string` as a default type value.