fix: current `color mode`
Saving the current `color mode` in a cookie to avoid `hydration issues`,
as this way the `server` already knows the current `color mode`.

It was necessary to use `useCookie()` to save the `color mode`
instead of the `storage setting` of `nuxtjs/color-mode`, as
the cookie value otherwise will be applied to the current `/path`.
Resulting in different `color modes` for different `paths.

See: https://github.com/nuxt-modules/color-mode/pull/301